combined firebase files successfully
